+++ main/actions/ledstripanimationactions.h
@@ -1,40 +1,14 @@
 #pragma once
 
 #include "actioninterface.h"
-#include "utils.h"
-#include "globals.h"
 #include "ledstrip.h"
 // #include "ledstripdefines.h"
 
-using namespace espgui;
-
-namespace {
 #ifdef FEATURE_LEDSTRIP
 template<uint16_t type>
-class LedStripSetAnimationAction : public virtual ActionInterface
+class LedStripSetAnimationAction : public virtual espgui::ActionInterface
 {
 public:
     void triggered() override { animation_type = type; }
 };
-
-/*
-class LedstripAnimationDefaultRainbowAction : public virtual ActionInterface
-{
-public:
-    void triggered() override { animation_type = LEDSTRIP_ANIMATION_TYPE_DEFAULTRAINBOW; }
-};
-
-class LedstripAnimationBetterRainbowAction : public virtual ActionInterface
-{
-public:
-    void triggered() override { animation_type = LEDSTRIP_ANIMATION_TYPE_BETTERRAINBOW; }
-};
-
-class LedstripAnimationSyncToSpeedAction : public virtual ActionInterface
-{
-public:
-    void triggered() override { animation_type = LEDSTRIP_ANIMATION_TYPE_SPEEDSYNCANIMATION; }
-};
-*/
 #endif
-}
