+++ main/displays/menus/ledstripselectotamode.h
@@ -2,20 +2,15 @@
 
 // Local includes
 #include "menudisplay.h"
-#include "utils.h"
-#include "menuitem.h"
-#include "ledstrip.h"
-#include "icons/back.h"
 #include "texts.h"
-#include "actions/switchscreenaction.h"
-#include "accessors/settingsaccessors.h"
-#include "ledstripmenu.h"
+#include "ledstrip.h"
+#include "globals.h"
+#include "utils.h"
 
 #if defined(FEATURE_LEDSTRIP) && defined(FEATURE_OTA)
-using namespace espgui;
 
 template <OtaAnimationModes mode>
-class LedstripChangeOtaAnimModeAction : public virtual ActionInterface
+class LedstripChangeOtaAnimModeAction : public virtual espgui::ActionInterface
 {
 public:
     void triggered() override
@@ -25,20 +20,12 @@ public:
     }
 };
 
-namespace {
-    class ledstripOtaAnimationChangeMenu :
-            public MenuDisplay,
-            public StaticText<TEXT_BLINKANIMATION>,
-            public BackActionInterface<SwitchScreenAction<LedstripMenu>>
-    {
-    public:
-        ledstripOtaAnimationChangeMenu()
-        {
-            constructMenuItem<makeComponent<MenuItem, StaticText<TEXT_OTAANIM_NONE>,          LedstripChangeOtaAnimModeAction<OtaAnimationModes::None>>>();
-            constructMenuItem<makeComponent<MenuItem, StaticText<TEXT_OTAANIM_PROGRESS>,      LedstripChangeOtaAnimModeAction<OtaAnimationModes::GreenProgressBar>>>();
-            constructMenuItem<makeComponent<MenuItem, StaticText<TEXT_OTAANIM_COLOR>,         LedstripChangeOtaAnimModeAction<OtaAnimationModes::ColorChangeAll>>>();
-            constructMenuItem<makeComponent<MenuItem, StaticText<TEXT_BACK>,                  SwitchScreenAction<LedstripMenu>, StaticMenuItemIcon<&espgui::icons::back>>>();
-        }
-    };
-} // Namespace
+class ledstripOtaAnimationChangeMenu :
+        public espgui::MenuDisplay,
+        public espgui::StaticText<TEXT_BLINKANIMATION>
+{
+public:
+    ledstripOtaAnimationChangeMenu();
+    void back() override;
+};
 #endif
